Module nscp makes some attempt to recognize and handle telnet IAC codes
This allows CTRL-C and CTRL-D to work from the client for people who
get stuck and don't want to use CTRL-].  It's a very dirt-simple
implementation of the telnet protocol so only standard Unix telnet is
supported.  The win32 telnet client sends too many IAC's that I do
not care to handle at this time, so they are just dropped with a
warning sent to the server.log.

(GetLine): password is no longer echoed back to the client.