Modified Ns_DriverInit to take new Ns_DriverInitData structure
Instead of passing all args as parameters to Ns_DriverInit, a
comm module must now create an Ns_DriverInitData structure and
populate it with appropriate values (see include/ns.h) and pass
that in the call to Ns_DriverInit. The Ns_DriverInitData
structure is versioned so that we can extend it later without
affecting other modules.