(NsTclICtlObjCmd): Fix intermittant core dump during oncleanup processing
The error is in the Tcl_Obj for the script passed into ns_ictl
on* callbacks was being saved and passed to the callbacks
(this Tcl_Obj was then being shared among Interps, which is
not allowed). Now, extract a copy of the script string, and
pass that to the callbacks.