Add ns_eval async completion
The early 4.0 ns_eval had been waiting for the completion of
integrating script changes into the server's init script.

This change adds the ability to return without waiting for the
init script integration completion (as the default). A new switch
modifies this behavior to force it to wait for the completion of
init script integration.

Added another switch to request a report of ns_eval backlog.