Specify MPL licence in headers; add cc to ns_sendmail; ns_getform returns "" when no conn
License terms in the headers are changed to Mozilla.

sendmail proc extended to support cc as argument and possibly ignore
error on receiving QUIT response, some mail servers just drop the
connection and then sendmail generate error.

ns_getform checks for connection and if no connection returns empty
instead of generate error, to be used safely in schedule scripts when
same code runs on the Web frontend and backend.