Remove ns_failservice, use SIGINT instead; remove experimental ReportException
Removed NsTclFailServiceObjCmd() and tied the processing to the NS_SIGINT
signal as sent by the -restart option of the "ns_shutdown" command, hence
avoiding different ways to force server restarts on different platforms.

Removed experimental ReportException() call since it does
not really seem to prevent the default handling in Windows where
a Window pops up and the user has to confirm (such behaviour
prevents the service control manager to restart the service).