Clarify responsibilities of Ns_CacheCreateEntry and Ns_CacheWaitCreateEntry
The first guarantees an Ns_Entry as a result and handles
expirey, the second can return NULL only on timeout waiting for
another thread to update the same entry.  Previously a
simultaneous flush would also cause a NULL return -- now however
the entry is recreated.

Fix bug with ExpireEntry: unset the value but do not delete the entry.