Pedantic additions of [list] quoting while constructing the script(s) used to initialize interp state
This is so that they won't blow up if someone were ever foolish
enough to create a namespace (or proc) that contains a space or
other meaningful Tcl punctuation ($, [], \, etc.)

Changed all instances of:
    uplevel <command> <args>
to explicitly be
    uplevel 1 <command> <args>
in case <command> were ever interpretable as an integer itself.

Changed "switch -glob $nop" to "switch -glob -- $nop" in case $nop were
to ever begin with a hyphen.