Add driver parameter 'acceptsize'
It is 1 by default, if set to higher number it will tell the driver to
accept more than one request, which as tests showed helps to process
more requests, because if we have many sockets in the driver's queue it takes
time to scan all of them and new requests are waiting to be accepted. Some
benchmarking showed that if set to 10 it improves req/sec rate by 10% on simple
ADP requests. If not set, driver works as usually.