Add config param schedsperthread, which causes thread to exit after n jobs
This is for schedule taks with -thread flags only, schedule tasks which are
executed inside the main sched thread still work as usuall.

This new config parameter similar to jobsperthread will allow to control threads
allocation and memory usage by exiting thread and reclaiming memory back to the
system in case of such memory allocator.