Fix for parsing nested <% ... %> in adp files
The previous code was counting how many levels deep the nesting of
<% ... %> tags was, but would only count down into the first
sequence of opening tags.  It is simpler to find the next closing tag
after the current closing tag each time an opening tag if found before
the current closing tag.  This handles an arbitrary nesting depth
and non-heirarchical nesting equally.