Add NS_EXPORT to modules
Recent versions of gcc include visibility attributes.  Tcl8.6
sets the '-fvisibility=hidden' flag when building, and naviserver
inherits this.  The effect is to keep all symbols with unspecified
visibility hidden in shared objects.  NS modules are shared objects
and need to export the symbols Ns_ModuleVersion and Ns_ModuleInit,
so these declarations need to have the NS_EXPORT declaration attached
to them.