add test for fastpath cache collision
The fastpath cache uses a filename as a key and verifies the freshness by
comparing modification times.  In the rare circumstance that the contents
of the file changes multiple times within the granularity of mtime (i.e.,
one second) a stale version of the file can be served out of the cache.